refactor($components): refactor first version codes
