+++ src/App.js
@@ -1,12 +1,11 @@
 import React, { Component } from 'react';
 import './App.css';
-import AddTodo from './containers/add_todo';
-import TodoList from './containers/todo_list';
+import AddTodo from './components/add_todo';
+import TodoList from './components/todo_list';
 
 class App extends Component {
     constructor(props) {
         super(props);
-
         this.state = {
             todos: []
         };
@@ -20,25 +19,27 @@ class App extends Component {
 
     handleDelete = (index, todos) => {
         todos.splice(index, 1);
-        console.log('todos:', todos);
+
         return todos;
+    };
+
+    deleteTodo = selectedTodoKey => {
+        this.setState({
+            todos: this.handleDelete(selectedTodoKey, this.state.todos)
+        });
     };
 
     handleEdit = (index, todos, todoTitle) => {
         todos[index] = todoTitle;
+
         return todos;
     };
 
-    editTodo(selectedTodoKey, todoTitle) {
-        console.log(`selectedTodoKey: ${selectedTodoKey}, todoTitle: ${todoTitle}`);
+    editTodo = (selectedTodoKey, todoTitle) => {
         this.setState({
-            todos: this.handleEdit(
-                selectedTodoKey,
-                this.state.todos,
-                todoTitle
-            )
+            todos: this.handleEdit(selectedTodoKey, this.state.todos, todoTitle)
         });
-    }
+    };
 
     render() {
         return (
@@ -46,15 +47,8 @@ class App extends Component {
                 <AddTodo onAddTodoChange={this.AddTodo} />
                 <TodoList
                     todos={this.state.todos}
-                    deleteTodo={selectedTodoKey => {
-                        this.setState({
-                            todos: this.handleDelete(
-                                selectedTodoKey,
-                                this.state.todos
-                            )
-                        });
-                    }}
-                    editTodo={this.editTodo.bind(this)}
+                    deleteTodo={this.deleteTodo}
+                    editTodo={this.editTodo}
                 />
             </div>
         );
